// convex/_generated/server.d.ts
type Env = {
  readonly SHOPIFY_API_VERSION: string | undefined;
  readonly SHOPIFY_CLIENT_ID: string | undefined;
  readonly SHOPIFY_CLIENT_SECRET: string | undefined;
  readonly SHOPIFY_STOREFRONT_ACCESS_TOKEN: string | undefined;
  readonly SHOPIFY_STOREFRONT_PUBLIC_ACCESS_TOKEN: string | undefined;
  readonly SHOPIFY_STORE_DOMAIN: string | undefined;
  readonly SHOPIFY_SYNC_SECRET: string | undefined;
};

/**
 * Define a query in this Convex app's public API.
 *
 * This function will be allowed to read your Convex database and will be accessible from the client.
 *
 * @param func - The query function. It receives a {@link QueryCtx} as its first argument.
 * @returns The wrapped query. Include this as an `export` to name it and make it accessible.
 */
export declare const query: QueryBuilder<DataModel, "public">;
// ...
export declare const action: ActionBuilder<DataModel, "public">;
// ...
export declare const env: Env;
// ...
export type ActionCtx = GenericActionCtx<DataModel>;